Extracted Applied_Versions class hierarchy
Migrating_Forward and Rolling_Backward are kind of like conditional
classes that abstract the note taking when a migration is (un)applied.

Applied_Versions deals with the intricacies of dealing with the database
schema table. I left the init_schema_table() method on Migrator because
the Migrator is the main entry point for the user.